set up template for react components
