+++ client/ListingEntry.jsx
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+
+const ListingEntry = (props) => {
+  return (
+    <div>
+    Listing entry here
+    </div>
+  )
+}
+export default ListingEntry;
