+++ client/index.jsx
@@ -1,21 +1,5 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
-
-class Listings extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      listings: [],
-    }
-  }
-
-  render() {
-    return (
-      <div>
-      Similar listings
-      </div>
-    )
-  }
-}
+import Listings from './Listings.jsx';
 
 ReactDOM.render(<Listings />, document.getElementById('similarlistings'))
